Moved to webpack 4
