--- =4.4.2
@@ -1,16 +1,0 @@
-
-> node-sass@4.11.0 install /home/den/projects/vue-crossword/node_modules/node-sass
-> node scripts/install.js
-
-Cached binary found at /home/den/.npm/node-sass/4.11.0/linux-x64-67_binding.node
-
-> node-sass@4.11.0 postinstall /home/den/projects/vue-crossword/node_modules/node-sass
-> node scripts/build.js
-
-Binary found at /home/den/projects/vue-crossword/node_modules/node-sass/vendor/linux-x64-67/binding.node
-Testing binary
-Binary is fine
-+ node-sass@4.11.0
-updated 1 package and audited 10068 packages in 6.678s
-found 2 vulnerabilities (1 low, 1 high)
-  run `npm audit fix` to fix them, or `npm audit` for details
