--- config/test.env.js
@@ -1,6 +1,0 @@
-const merge = require('webpack-merge')
-const devEnv = require('./dev.env')
-
-module.exports = merge(devEnv, {
-  NODE_ENV: '"testing"',
-})
