+++ package.json
@@ -1,81 +1,36 @@
 {
-  "name": "vue-crossword-app",
-  "description": "Some drafts about crossword app.",
-  "version": "0.0.283",
+  "name": "boiler",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint"
+  },
   "dependencies": {
-    "11": "0.0.1",
     "axios": "^0.18.0",
-    "oauth": "^0.9.15",
-    "vue": "^2.6.7",
-    "vue-router": "3.0.1"
+    "core-js": "^2.6.5",
+    "vue": "^2.6.10"
   },
   "devDependencies": {
-    "autoprefixer": "^7.2.6",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-istanbul": "^5.1.2",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.22.0",
-    "chromedriver": "^2.46.0",
-    "connect-history-api-fallback": "^1.6.0",
-    "copy-webpack-plugin": "^4.6.0",
-    "cross-env": "^5.2.0",
-    "cross-spawn": "^5.0.1",
-    "css-loader": "^2.1.1",
-    "cssnano": "^4.1.10",
-    "eslint": "^5.14.1",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-loader": "^2.1.2",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-node": "^8.0.1",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^5.2.2",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.16.4",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "file-loader": "^0.11.1",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "html-webpack-plugin": "^2.28.0",
-    "http-proxy-middleware": "^0.19.1",
-    "inject-loader": "^3.0.0",
-    "opn": "^5.4.0",
-    "optimize-css-assets-webpack-plugin": "^2.0.0",
-    "ora": "^1.4.0",
-    "phantomjs-prebuilt": "^2.1.14",
-    "rimraf": "^2.6.3",
-    "sass-loader": "^6.0.7",
-    "semver": "^5.6.0",
-    "shelljs": "^0.7.6",
+    "@vue/cli-plugin-babel": "^3.7.0",
+    "@vue/cli-plugin-eslint": "^3.7.0",
+    "@vue/cli-service": "^3.7.0",
+    "babel-eslint": "^10.0.1",
+    "cache-loader": "^3.0.0",
+    "eslint": "^5.16.0",
+    "eslint-plugin-vue": "^5.0.0",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.2",
-    "ts-loader": "^2.3.3",
-    "vue-eslint-parser": "^3.3.0",
-    "vue-loader": "^12.1.0",
-    "vue-style-loader": "^3.1.2",
-    "vue-template-compiler": "^2.6.7",
-    "webpack": "^2.6.1",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-dev-middleware": "^1.10.0",
-    "webpack-hot-middleware": "^2.24.3",
-    "webpack-merge": "^4.2.1"
+    "vue-template-compiler": "^2.5.21"
   },
-  "scripts": {
-    "dev": "node build/dev-server.js",
-    "start": "node build/dev-server.js",
-    "build": "node build/build.js",
-    "lint": "eslint --ext .js,.vue src"
-  },
-  "engines": {
-    "node": ">= 9.0.0",
-    "npm": ">= 5.0.0"
+  "postcss": {
+    "plugins": {
+      "autoprefixer": {}
+    }
   },
   "browserslist": [
     "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
+    "last 2 versions"
   ]
 }
