+++ src/components/WordForm.vue
@@ -1,12 +1,12 @@
 <template>
   <div
-    @mouseover="$emit('form-hovered', { x, y, isVertical, length })"
-    @mouseleave="$emit('form-leaved')"
     :class="{
       next: nextQuery === query,
       error: suggestionsCount === 0,
     }"
     class="word-form"
+    @mouseover="$emit('form-hovered', { x, y, isVertical, length })"
+    @mouseleave="$emit('form-leaved')"
   >
     <div v-if="loading && !suggested.length" class="word-form-overlay">
       <div class="inner">
@@ -16,7 +16,7 @@
 
     <div v-if="suggestionsVisible" class="modal-area suggestions">
       <div class="inner">
-        <a @click.prevent="hideSuggestions" href="#" class="close">
+        <a href="#" class="close" @click.prevent="hideSuggestions">
           X
         </a>
         <ul class="suggested-list">
@@ -26,17 +26,17 @@
             :style="{ width: `${length / 1.618}em` }"
           >
             <a
+              href="#"
               @mouseover="$emit('suggestion-hover', { x, y, isVertical, word })"
               @mouseleave="$emit('suggestion-leave')"
               @click.prevent="pasteWord(wordItem)"
               v-html="getWordHtml(wordItem)"
-              href="#"
             ></a>
           </li>
         </ul>
         <div class="controls">
           <div class="prev go-to-page">
-            <a @click.prevent="prevPage" href="#">
+            <a href="#" @click.prevent="prevPage">
               Prev.
             </a>
           </div>
@@ -44,7 +44,7 @@
             Page {{ page + 1 }}
           </div>
           <div class="next go-to-page">
-            <a @click.prevent="nextPage" href="#">
+            <a href="#" @click.prevent="nextPage">
               Next.
             </a>
           </div>
@@ -54,15 +54,15 @@
 
     <div v-if="cluesVisible" class="modal-area clues">
       <div class="inner">
-        <a @click.prevent="hideClues" href="#" class="close">
+        <a href="#" class="close" @click.prevent="hideClues">
           X
         </a>
         <ol class="clues-list">
           <li v-for="clue in ownClues" :key="clue.id">
             <a
+              href="#"
               @click.prevent="pasteClue(clue)"
               v-html="getClueHtml(clue)"
-              href="#"
             ></a>
             <br />
           </li>
@@ -85,27 +85,27 @@
           <input
             v-model="answer[idx]"
             :data-idx="idx"
-            @blur="onBlur"
-            @focus="onFocus"
-            @input="onInputLetter"
-            @paste.prevent.stop="onPaste"
             type="text"
             size="1"
             minlength="1"
             maxlength="1"
+            @blur="onBlur"
+            @focus="onFocus"
+            @input="onInputLetter"
+            @paste.prevent.stop="onPaste"
           />
         </div>
       </div>
       <div class="clue">
         <textarea
-          ref="question"
           v-if="word"
+          ref="question"
           v-model="question"
           :cols="cols"
           :rows="rows"
-          @focus="showClues"
           type="text"
           class="textarea"
+          @focus="showClues"
         ></textarea>
       </div>
       <div v-if="noClue" class="green-plus">
@@ -116,17 +116,17 @@
 
     <div class="words-count">
       <a
-        @click.prevent="toggleModal(query)"
-        v-text="suggestionsText"
         href="#"
         title="Show variants"
+        @click.prevent="toggleModal(query)"
+        v-text="suggestionsText"
       ></a>
       <a
         v-if="word"
-        @click.prevent="removeWord"
         class="danger"
         href="#"
         title="Remove word"
+        @click.prevent="removeWord"
       >
         Remove word
       </a>
