+++ vue.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  runtimeCompiler: true,
+}
